Extract scmsync related methods for workflow steps to a concern
