+++ src/api/app/models/workflow/step/branch_package_step.rb
@@ -1,4 +1,6 @@
 class Workflow::Step::BranchPackageStep < ::Workflow::Step
+  include ScmSyncEnabledStep
+
   REQUIRED_KEYS = [:source_project, :source_package, :target_project].freeze
   validate :validate_source_project_and_package_name
